+++ src/main/java/ir/tic/clouddc/center/CenterService.java
@@ -9,9 +9,9 @@ import java.util.List;
 public interface CenterService {
 
     Model getCenterLandingPageModel(Model model);
-    Salon getSalon(long salonId);
-    Location getLocation(long locationId);
-    Model getLocationDetailModel(long locationId, Model model);
+    Salon getSalon(int salonId);
+    Location getLocation(int locationId);
+    Model getLocationDetailModel(int locationId, Model model);
     Center getCenter(int centerId);
     List<Salon> getSalonList();
     List<Center> getCenterList();
